If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,6 @@
 #include "function_pointers.h"
 #include <stdlib.h>
+
 /**
  * array_iterator - executes a function on param of each array element.
  * @array: an array
@@ -15,6 +16,6 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 	if (array == NULL || action == NULL)
 		return;
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int) size; i++)
 		action(array[i]);
 }
